Further refactoring to bring closer to PSR standards
