+++ mangapress.php
@@ -55,7 +55,39 @@ if (!defined('MP_LANG'))
 if (!defined('MP_DOMAIN'))
     define('MP_DOMAIN', 'mangapress');
 
-require_once MP_ABSPATH . 'includes/lib/form/element.php';
+
+/**
+ * Autoloader
+ */
+spl_autoload_register(function ($class) {
+
+    // project-specific namespace prefix
+    $prefix = 'MangaPress\\';
+
+    // base directory for the namespace prefix
+    $base_dir = __DIR__ . '/includes/lib/';
+
+    // does the class use the namespace prefix?
+    $len = strlen($prefix);
+    if (strncmp($prefix, $class, $len) !== 0) {
+        // no, move to the next registered autoloader
+        return;
+    }
+
+    // get the relative class name
+    $relative_class = substr($class, $len);
+
+    // replace the namespace prefix with the base directory, replace namespace
+    // separators with directory separators in the relative class name, append
+    // with .php
+    $file = $base_dir . str_replace('\\', '/', strtolower($relative_class)) . '.php';
+
+    // if the file exists, require it
+    if (file_exists($file)) {
+        require $file;
+    }
+});
+
 require_once MP_ABSPATH . 'includes/lib/class-flash-messages.php';
 require_once MP_ABSPATH . 'includes/lib/class-content-type.php';
 require_once MP_ABSPATH . 'includes/lib/class-post-type.php';
